feat(Core/Support): 增加部分核心/支撑功能模块
1. 增加异常模块koala-exceptions
2. 增加工具模块koala-utils
3. 增加MyBatis支撑模块koala-support-mybatis
4. 增加存储库MyBatis实现模块koala-support-repositories-mybatis

+++ koala-core/koala-models/src/main/java/cn/houtaroy/koala/models/Sortable.java
@@ -10,5 +10,5 @@ public interface Sortable {
    *
    * @return 排序索引
    */
-  Long getSortIndex();
+  Integer getSortIndex();
 }
